CreditCheckAgent/requirements.txt
strands-agents
a2a-sdk
uvicorn
fastapi
anthropic
ollama
